Switch to LLVM C++ formatting
We used a weird mix of FreeBSD style recommendations and extremely long lines. This made the code unreadable after some time. There are multiple code guidelines from various big projects but since LLVM is used to complite this code on FreeBSD, we can as well use it to style code.

You might want to configure your editor to re-style files when saving them.